+++ Kodumun Yeri/Gülücük.c
@@ -1,0 +1,11 @@
+#include <stdio.h>
+
+int main (){
+	
+	
+	
+	char a[8]="sefa              kalkan"; // gülücük çýkaran kod
+	
+	printf(a);
+	
+}
