+++ Week 3/ödev-1.c
@@ -1,0 +1,5 @@
+#include <stdio.h>
+
+int main (){
+	printf("     * \n    * *\n   *   *\n  *     * \n ********* ");	
+}
