+++ Week-10/TekCift.c
