--- Ödev/Serbest calisma/Gülücük.c
@@ -1,11 +1,0 @@
-#include <stdio.h>
-
-int main (){
-	
-	
-	
-	char a[8]="sefa              kalkan"; // gülücük çýkaran kod
-	
-	printf(a);
-	
-}
