--- Ödev/week 2/ödev-3.c
@@ -1,11 +1,0 @@
-#include <stdio.h>
-
-int main (){
-	char a,b;
-	printf("Lütfen isminizin baþ harflerini giriniz : ");
-	scanf("%c %c", a,b);
-	
-	printf("Hoþ geldiniz %c %c",a,b);
-	return 0;
-}
-
