--- Ödev/week 3/ödev-1.c
@@ -1,5 +1,0 @@
-#include <stdio.h>
-
-int main (){
-	printf("     * \n    * *\n   *   *\n  *     * \n ********* ");	
-}
