--- Ödev/week 4/ödev-1.c
@@ -1,18 +1,0 @@
-#include <stdio.h>
-
-int main () {
-	int a;
-	
-	printf("Sayinizi giriniz : ");
-	scanf("%d", &a);
-	
-	if (a%2 == 0){	
-	printf("Sayiniz cifttir");
-	}
-	
-	else {	
-	printf("Sayiniz tektir");
-	} 
-	
-	return 0;
-}
